embedding_layer/requirements.txt
boto3==1.34.34
langchain==0.1.9
nltk==3.8.1
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
sentence-transformers==2.5.1 